start implementing cache views

+++ src/main/java/com/mewna/catnip/entity/impl/GuildImpl.java
@@ -88,7 +88,7 @@ public class GuildImpl implements Guild, RequiresCatnip, Timestamped {
     private String joinedAt;
     private boolean large;
     private boolean unavailable;
-    private int memberCount;
+    private long memberCount;
     
     @Override
     public void catnip(@Nonnull final Catnip catnip) {
